#204: moved common piecewise linear function code to a plugin lib
