+++ script/VeLa/Y/Y.vela
@@ -3,7 +3,7 @@
 # nor are function parameter and return types fully
 # specified (just "function") yet
 
-Y is λ(h : function) : function {
+Y is λ(h : λ) : function {
     λ(f : function) : function {
         f(f)
     } (λ(f : function) : function {
